# Conflicts:
#	ruoyi-modules/xanadu-customer-center/src/main/java/edu/neu/cc/controller/RefundController.java
#	ruoyi-modules/xanadu-distribution-center/src/main/java/edu/neu/dbc/controller/SupplierController.java

+++ ruoyi-modules/xanadu-distribution-center/src/main/java/edu/neu/dbc/controller/SupplierController.java
@@ -1,10 +1,13 @@
 package edu.neu.dbc.controller;
 
 
+import com.baomidou.mybatisplus.core.conditions.query.QueryWrapper;
 import com.baomidou.mybatisplus.extension.plugins.pagination.Page;
 import com.baomidou.mybatisplus.core.conditions.query.QueryWrapper;
 import com.ruoyi.common.core.web.domain.AjaxResult;
+import edu.neu.dbc.entity.PurchaseRecord;
 import edu.neu.dbc.entity.Supplier;
+import edu.neu.dbc.service.PurchaseRecordService;
 import edu.neu.dbc.service.SupplierService;
 import io.swagger.annotations.ApiOperation;
 import org.springframework.beans.factory.annotation.Autowired;
@@ -15,7 +18,7 @@ import java.util.Map;
 
 /**
  * <p>
- *  前端控制器
+ * 前端控制器
  * </p>
  *
  * @author Gaosong Xu
@@ -30,6 +33,9 @@ public class SupplierController {
 
     @Autowired
     SupplierService supplierService;
+
+    @Autowired
+    PurchaseRecordService purchaseRecordService;
 
 
     @GetMapping("/list")
@@ -67,7 +73,10 @@ public class SupplierController {
     @DeleteMapping("/delete/{id}")
     @ApiOperation("删除供应商")
     public AjaxResult delete(@PathVariable Long id) {
-        //TODO: 检查供应商是否有过采购记录
+        QueryWrapper<PurchaseRecord> purchaseRecordQueryWrapper = new QueryWrapper<PurchaseRecord>().eq("supplier_id", id);
+        if (purchaseRecordService.count(purchaseRecordQueryWrapper) > 0) {
+            return AjaxResult.error("该供应商有过采购记录，无法删除");
+        }
         return AjaxResult.success(supplierService.removeById(id));
     }
